+++ src/shared/Router.jsx
@@ -18,9 +18,8 @@ const Router = () => {
                 <Route path="" element={<Layout />}>
                     <Route path="/" element={<Home />} />
                     <Route path="/detail" element={<Detail />} />
-                    <Route path="/profile" element={<Profile />} />
-                    <Route path="/profile/*" element={<Profile />} />
-                    {/* <Route path="/profile/:id/*" element={<Profile />} /> */}
+                    <Route path="/profile/:userId" element={<Profile />} />
+                    <Route path="/profile/:userId/*" element={<Profile />} />
                     <Route path="/search" element={<Search />} />
                     <Route path="/form" element={<Form />} />
                     <Route path="/modify-profile" element={<ModifyProfile />} />
